+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.5.0 / Unreleased
 
-* 
+* [CHANGE] Changed widget namespace from 'ui' to 'moogle'.
 
 # 0.4.0 / 2013-05-28